:recycle: :: (#2) Change NotificationSender function parameter type
