+++ src/main/kotlin/App.kt
@@ -1,18 +1,25 @@
 import com.slack.api.bolt.App
 import com.slack.api.bolt.socket_mode.SocketModeApp
-import handler.AmsHandler
+import handler.MeetingHandler
 import sender.notification.NotificationMessageSender
 
 fun main() {
     val app = App()
     val notifySender = NotificationMessageSender()
-    val amsHandler = AmsHandler(notifySender)
+    val meetingHandler = MeetingHandler(notifySender)
 
-    app.command("/meeting", amsHandler::addSchedule)
-    app.viewSubmission("add-schedule", amsHandler::handleViewSchedule)
+    // Commands
+    app.command("/meeting", meetingHandler::addSchedule)
+
+    // Modal
+    app.viewSubmission("add-schedule", meetingHandler::handleViewSchedule)
     app.viewClosed("add-schedule") { _, ctx ->
         ctx.ack()
     }
+
+    // Actions
+    app.blockAction("approve-meeting", meetingHandler::handleApproveAction)
+    app.blockAction("deny-meeting", meetingHandler::handleDenyAction)
 
     SocketModeApp(app).start()
 }
