+++ src/main/kotlin/sender/dto/Meeting.kt
@@ -1,0 +1,10 @@
+package sender.dto
+
+class Meeting (
+    val agenda: String,
+    val attender: List<String>,
+    val description: String,
+    val date: String,
+    val hour: String,
+    val minute: String
+)
